Changes to UserController class
Had to make changes to the user controller to allow the creation of isolated instances for the unit tests. Changes made: allow the class to be exported, added a optional parameter for services in the contructor, added an method to return an isolated instance for unit test, to no mess with the private constructor.